Reorganização dos arquivos, componentes e styled
