+++ semana9/labex/src/components/CardAdminDetails.js
@@ -1,20 +1,55 @@
-import React from "react";
 import axios from "axios";
-import { headers } from "../api/apiUtils";
+import React from "react";
+import { useHistory } from "react-router";
+import { goToCreteTripPage, goToTripDetailsPage } from "../routes/coordinator";
+import { headers, URL_TRIPS } from "../utils/apiUtils";
 
-const CardAdminDetails = () => {
-  const decideApproved = (tripId, candidateId, choice) => {
-    const body = {
-      approve: choice,
-    };
-    axios.put(
-      `https://us-central1-labenu-apis.cloudfunctions.net/labeX/angelo-odwyer-cruz/trips/${tripId}/candidates/${candidateId}/decide`,
-      body,
-      headers
-    );
-  };
 
-  return <></>;
+const CardAdminDetails = (props) => {
+  const history = useHistory()
+  
+  const data = props.data
+  
+  const logout = props.logout
+
+  const deleteTrip = (id) => {
+
+    axios.delete(`${URL_TRIPS}/${id}`, headers)
+    .then(res =>{
+      alert("Viagem deletada com sucesso!")
+      window.location.reload()
+    })
+    .catch(err => {
+      console.log(err)
+    })
+  }
+
+  return (
+  <div>
+    <h1>Página do Administrador</h1>
+          <button onClick={logout}>Logout</button>
+          <button onClick={() => goToCreteTripPage(history)}>
+            Criar Viagem
+          </button>
+          {data.trips &&
+            data.trips.map((trip) => {
+              return (
+                <div>
+                  <p>
+                    {trip.name}
+                  </p>
+                  <button  onClick={() => goToTripDetailsPage(history, trip.id)}>
+                    Detalhes da Viagem
+                  </button>
+                  
+                  <button  onClick={() => deleteTrip(trip.id)}>
+                    Deletar da Viagem
+                  </button>
+                </div>
+              );
+            })}
+  </div>
+  );
 };
 
 export default CardAdminDetails;
