+++ semana9/labex/src/routes/Router.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
-import AdiminHomePage from "../pages/AdminHome";
-import ApplyFormsPage from "../pages/ApplyForm";
-import CreateTripPage from "../pages/CreateTrip";
-import HomePage from "../pages/Home";
-import ListTripPage from "../pages/ListTrip";
-import LoginPage from "../pages/Login";
-import TripDetailsPage from "../pages/TtripDetails";
+import AdiminHomePage from "../pages/AdminHomePage/AdminHome";
+import ApplyFormsPage from "../pages/ApplyFormPage/ApplyForm";
+import CreateTripPage from "../pages/CreateTripPage/CreateTrip";
+import HomePage from "../pages/HomePage/Home";
+import ListTripPage from "../pages/ListTripPage/ListTrip";
+import LoginPage from "../pages/LoginPage/Login";
+import TripDetailsPage from "../pages/TripDetailsPage/TripDetails";
 
 const Router = () => {
   return (
